\name{RSmap-package}
\alias{RSmap-package}
\docType{package}
\title{
R client bindings for sMAP archive
}
\description{
Connect to and query a sMAP archive using the client API.
}
\details{
\tabular{ll}{
Package: \tab RSmap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-09\cr
License: \tab GPL\cr
}
}
\author{
Tyler Hoyt <tyler.hoyt@gmail.com>
}
\references{
http://pythonhosted.org/Smap/
}
\keyword{ package }
